agents/chart-maker/pyproject.toml
[project]
name = "chart_maker"
version = "0.1.0"
description = "A chart maker agent"
authors = [
    {name = "Weiyuan Liu"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-adk (>=1.17.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "a2a-sdk (>=0.3.10,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
